+++ client/changelog.md
@@ -1,1 +1,2 @@
 - Fixed crash when entering singeplayer twice
+- Fixed crash when trying to change hud icon/overlay position